# param add perf counters