--- a/src/modules/commander/arm_auth.cpp
+++ b/src/modules/commander/arm_auth.cpp
@@ -179,9 +179,11 @@ uint8_t arm_auth_check()
 	return vehicle_command_ack_s::VEHICLE_RESULT_DENIED;
 }
 
-void arm_auth_update(hrt_abstime now)
+void arm_auth_update(hrt_abstime now, bool param_update)
 {
-	param_get(param_arm_parameters, (int32_t*)&arm_parameters);
+	if (param_update) {
+		param_get(param_arm_parameters, (int32_t*)&arm_parameters);
+	}
 
 	switch (state) {
 	case ARM_AUTH_WAITING_AUTH: