--- a/src/modules/commander/arm_auth.h
+++ b/src/modules/commander/arm_auth.h
@@ -45,6 +45,6 @@ enum arm_auth_methods {
 };
 
 void arm_auth_init(orb_advert_t *mav_log_pub, uint8_t *system_id);
-void arm_auth_update(hrt_abstime now);
+void arm_auth_update(hrt_abstime now, bool param_update = true);
 uint8_t arm_auth_check();
 enum arm_auth_methods arm_auth_method_get();