--- a/src/modules/commander/commander.cpp
+++ b/src/modules/commander/commander.cpp
@@ -1704,9 +1704,10 @@ Commander::run()
 		arming_ret = TRANSITION_NOT_CHANGED;
 
 		/* update parameters */
-		orb_check(param_changed_sub, &updated);
+		bool params_updated = false;
+		orb_check(param_changed_sub, &params_updated);
 
-		if (updated || param_init_forced) {
+		if (params_updated || param_init_forced) {
 
 			/* parameters changed */
 			struct parameter_update_s param_changed;
@@ -3189,7 +3190,7 @@ Commander::run()
 			have_taken_off_since_arming = false;
 		}
 
-		arm_auth_update(now);
+		arm_auth_update(now, params_updated || param_init_forced);
 
 		usleep(COMMANDER_MONITORING_INTERVAL);
 	}