--- a/src/modules/navigator/follow_target.cpp
+++ b/src/modules/navigator/follow_target.cpp
@@ -80,7 +80,6 @@ FollowTarget::FollowTarget(Navigator *navigator, const char *name) :
 	_yaw_auto_max(0.0F),
 	_yaw_angle(0.0F)
 {
-	updateParams();
 	_current_target_motion = {};
 	_previous_target_motion =  {};
 	_current_vel.zero();