--- a/src/modules/navigator/takeoff.cpp
+++ b/src/modules/navigator/takeoff.cpp
@@ -57,8 +57,6 @@ Takeoff::Takeoff(Navigator *navigator, const char *name) :
 	MissionBlock(navigator, name),
 	_param_min_alt(this, "MIS_TAKEOFF_ALT", false)
 {
-	// load initial params
-	updateParams();
 }
 
 Takeoff::~Takeoff()