… src/modules/sensors/temp_comp_params_gyro.c
… src/modules/sensors/temperature_compensation.cpp