nsh> perf
	...
	param_set: 0 events, 0us elapsed, 0us avg, min 0us max 0us 0.000us rms
	param_get: 5266 events, 55534us elapsed, 10us avg, min 4us max 1893us 57.558us rms
	param_export: 0 events, 0us elapsed, 0us avg, min 0us max 0us 0.000us rms

	nsh> param set SYS_AUTOSTART 2100
	nsh> param save


	nsh> perf
	...
	param_set: 0 events, 0us elapsed, 0us avg, min 0us max 0us 0.000us rms
	param_get: 6960 events, 78366us elapsed, 11us avg, min 4us max 1896us 61.081us rms
	param_export: 1 events, 1012771us elapsed, 1012771us avg, min 1012771us max 1012771us   NaNus rms


	nsh> param set SYS_MC_EST_GROUP 1
	  SYS_MC_EST_GROUP: curr: 2 -> new: 1
	nsh> param save


	nsh> perf
	...
	param_set: 1 events, 86us elapsed, 86us avg, min 86us max 86us Infinityus rms
	param_get: 18022 events, 196162us elapsed, 10us avg, min 4us max 2690us 62.364us rms
	param_export: 3 events, 3288253us elapsed, 1096084us avg, min 1012771us max 1247173us 131077.047us rms