@bkueng 2017-11-30
I'm wondering if it's worth adding the overhead to `param_get`, since it's a fast method (thus adding `perf_end` will have a noticeable impact) and it's called many times on a param change event.
Do we know what the cost is for a perf counter?