@LorenzMeier 2017-11-30
The overhead is not insignificant, that’s correct. Should this be a compile time option for benchmarks?