@dagar 2017-11-30
That could be fine, I mainly want to use this to find the abusers. There shouldn't be anywhere near this many calls to param_get().