@dagar 2017-12-11
On second thought I'd like to do some testing to see if we can tolerate leaving this on to catch misuse in regular operation. There are also more areas of the system where we shouldn't even be updating params while flying.