@dagar 2017-12-12
@bkueng could you give this an initial review (the last commit in particular). It drastically reduces the param export time on a pixhawk. The param export is called from either the param command `param save` or in LPWORK. In both cases there's plenty of stack available for the buffer.

Pixhawk 2 baseline

    param_export: 3 events, 3288253us elapsed, 1096084us avg, min 1012771us max 1247173us 131077.047us rms

Pixhawk 2 with bson encoder buffer

    param_export: 3 events, 30897us elapsed, 10299us avg, min 10219us max 10349us 70.004us rms