@bkueng 2017-12-12
Nice! I just glanced over it but it makes sense. So it seems that NuttX does not use a write buffer for the params file. I have a few questions/remarks:
- What is the contribution of `usleep(5);` vs. using the buffer alone?
- is there any change in error behavior when an export is interrupted? e.g. a power loss.
- there are a few more callers through `param_save_default()` in the codebase