@dagar 2017-12-12
> What is the contribution of usleep(5); vs. using the buffer alone?
I'm not sure exactly (it was dropped later on). I can do a quick test with and without. Do you see any reason to keep it? I believe all the potentially callers are low priority.

> is there any change in error behavior when an export is interrupted? e.g. a power loss.
Change yes, but I think either way you're screwed and end up with partial bson records and no bson end. If this is a major concern I have some other (initially unrelated) ideas for the param system that would actually help here.

> there are a few more callers through param_save_default() in the codebase
Thanks, I'll take a closer look. Even with a smaller buffer it's still worth it.


I've been testing across a random assortment of boards. Obviously the timing is dependant on the number of params, but curiously the pixhawk 2 cube seems to take much longer to save (often > 1s in master) compared to a pixracer (usually < 500ms).