@bkueng 2017-12-13
> I can do a quick test with and without. Do you see any reason to keep it?

No, I just like to know the individual contributing factors.

> If this is a major concern I have some other (initially unrelated) ideas for the param system that would actually help here.

Ok. So do I, but I think we're fine. Since it's much faster now, it's becoming less and less of a concern. And since the plan is for future boards to have a small backup battery, it will become easier to handle as well.

Did you see the NuttX options `CONFIG_MTD_WRBUFFER`, `CONFIG_MTD_NWRBLOCKS` and `CONFIG_RAMTRON_WRITEWAIT` (fyi @davids5 )?