@dagar 2018-01-04
I haven't looked experimented with the NuttX write buffers yet, but I'm wondering if we should go with this regardless. Even on LInux (desktop x86 or Raspi Arm) it's much faster.

master + perf
	param_export: 6 events, 15519us elapsed, 2586us avg, min 2281us max 3157us 318.050us rms

PR with bson encoder buffer
	param_export: 5 events, 116us elapsed, 23us avg, min 0us max 45us 21.534us rms